pyproject.toml
[project]
name = "chat-with-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "faiss-cpu",
    "langchain==0.3",
    "langchain-community",
    "langchain-google-genai",
    "pydantic",
    "pypdf",
    "python-dotenv",
    "python-multipart",
    "redis",
    "uvicorn",
    "cachetools",
    "slowapi",

]
